embedding-dump: default to f16 and add q8_0 storage
The dump is the bulk of the disk cost for a training corpus, and the previous
default was f32. Make f16 the default - it halves the size at a negligible cost
for activations - and add q8_0 for corpora that do not fit.

q8_0 uses ggml's own block format, so the values are bit-identical to what a
q8_0 KV cache stores. numpy has no block-quantized dtype, so each array is
written as an int8 payload plus its per-32-value f16 scales in '<name>_scales';
dequantize with `qs * scales.repeat(32, axis=-1)`. The array is rejected if its
row width is not a multiple of 32.

Measured on the test model, per token: f32 2326 B, f16 1174 B, q8_0 640 B - so
q8_0 is 1.83x smaller than f16 and 3.6x smaller than f32.

Tests/test-embedding-dump-kv.py now pins the default to f16 and checks that the
q8_0 payload dequantizes back to the f32 values, in addition to the existing
row-alignment and post-rope checks.

+++ common/common.h
@@ -728,7 +728,7 @@ struct common_params {
                                        // "roles" (+roles int32), "turns" (+roles, turn_id)
     std::string dump_kv_layers = "";   // per-layer KV rows to dump: "" (none), a comma-separated
                                        // list of layer indices, or "all"
-    std::string dump_dtype = "f32";    // on-disk dtype for the hidden/KV arrays: "f32" or "f16"
+    std::string dump_dtype = "f16";    // on-disk dtype for the hidden/KV arrays: "f32", "f16" or "q8_0"
     bool        dump_hidden = true;    // dump the target hidden state (llama_get_embeddings_nextn)
 
     // server params
